requirements-dev.txt
# Development and testing dependencies
# Install with: pip install -r requirements-dev.txt

# Security auditing
pip-audit>=2.6.0
safety>=2.3.0

# Code quality
flake8>=6.1.0
pylint>=3.0.0
black>=23.12.0
isort>=5.13.0

# Testing
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
coverage>=7.3.0

# Django testing utilities
django-test-plus>=1.4.0
model-bakery>=0.3.0

